Narrow character level type from u32 to u16

+++ src/scenes/character_frame/mod.rs
@@ -134,9 +134,9 @@ fn build_state(
     }
 }
 
-fn extract_identity(stats: Option<&CharacterStatsSnapshot>) -> (String, u32, String) {
+fn extract_identity(stats: Option<&CharacterStatsSnapshot>) -> (String, u16, String) {
     let name = stats.and_then(|s| s.name.clone()).unwrap_or_default();
-    let level = stats.and_then(|s| s.level).unwrap_or(0);
+    let level = stats.and_then(|s| s.level).unwrap_or(0) as u16;
     let class = stats
         .and_then(|s| s.class)
         .map(class_name_from_id)
